+++ database/sql/src/main/scala/cromwell/database/sql/joins/CallCachingDiffJoin.scala
@@ -1,0 +1,5 @@
+package cromwell.database.sql.joins
+
+import cromwell.database.sql.tables.CallCachingEntry
+
+case class CallCachingDiffJoin(cacheEntryA: CallCachingEntry, cacheEntryB: CallCachingEntry, diff: Seq[(Option[(String, String)], Option[(String, String)])])