+++ engine/src/main/scala/cromwell/webservice/WorkflowJsonSupport.scala
@@ -10,7 +10,7 @@ import MetadataService.{WorkflowQueryResponse, WorkflowQueryResult}
 import cromwell.util.JsonFormatting.WdlValueJsonFormatter
 import WdlValueJsonFormatter._
 import better.files.File
-import spray.json.{DefaultJsonProtocol, JsString, JsValue, RootJsonFormat}
+import spray.json._
 
 object WorkflowJsonSupport extends DefaultJsonProtocol {
   implicit val workflowStatusResponseProtocol = jsonFormat2(WorkflowStatusResponse)