+++ engine/src/main/scala/cromwell/webservice/metadata/MetadataBuilderActor.scala
@@ -1,11 +1,7 @@
 package cromwell.webservice.metadata
 
-import java.time.OffsetDateTime
-
 import akka.actor.{ActorRef, LoggingFSM, Props}
 import cromwell.webservice.metadata.MetadataComponent._
-import cats.instances.list._
-import cats.syntax.foldable._
 import cromwell.core.Dispatcher.ApiDispatcher
 import cromwell.core.ExecutionIndex.ExecutionIndex
 import cromwell.core.{WorkflowId, WorkflowMetadataKeys, WorkflowState}
@@ -20,11 +16,6 @@ import spray.http.{StatusCodes, Uri}
 import spray.httpx.SprayJsonSupport._
 import spray.json._
 
-import scala.collection.immutable.TreeMap
-import scala.language.postfixOps
-import scala.util.{Failure, Random, Success, Try}
-
-
 object MetadataBuilderActor {
   sealed trait MetadataBuilderActorState
   case object Idle extends MetadataBuilderActorState
@@ -32,95 +23,26 @@ object MetadataBuilderActor {
   case object WaitingForSubWorkflows extends MetadataBuilderActorState
 
   case class MetadataBuilderActorData(
-                                      originalQuery: MetadataQuery,
-                                      originalEvents: Seq[MetadataEvent],
-                                      subWorkflowsMetadata: Map[String, JsValue],
-                                      waitFor: Int
-                                    ) {
+                                       originalQuery: MetadataQuery,
+                                       originalEvents: Seq[MetadataEvent],
+                                       subWorkflowsMetadata: Map[String, JsValue],
+                                       waitFor: Int
+                                     ) {
     def withSubWorkflow(id: String, metadata: JsValue) = {
       this.copy(subWorkflowsMetadata = subWorkflowsMetadata + ((id, metadata)))
     }
-    
+
     def isComplete = subWorkflowsMetadata.size == waitFor
   }
-  
+
   def props(serviceRegistryActor: ActorRef) = {
     Props(new MetadataBuilderActor(serviceRegistryActor)).withDispatcher(ApiDispatcher)
   }
 
   val log = LoggerFactory.getLogger("MetadataBuilder")
 
-  private val KeySeparator = MetadataKey.KeySeparator
-  private val bracketMatcher = """\[(\d*)\]""".r
   private val AttemptKey = "attempt"
   private val ShardKey = "shardIndex"
-
-  def parseKeyChunk(chunk: String, innerValue: MetadataComponent): MetadataComponent = {
-    chunk.indexOf('[') match {
-      // If there's no bracket, it's an object. e.g.: "calls"
-      case -1 => MetadataObject(Map(chunk -> innerValue))
-      // If there's a bracket it's a named list. e.g.: "executionEvents[0][1]"
-      case bracketIndex =>
-        // Name: "executionEvents"
-        val objectName = chunk.substring(0, bracketIndex)
-        
-        // Empty value means empty list
-        if (innerValue == MetadataEmpty) MetadataObject(Map(objectName -> MetadataList(Map.empty)))
-        else {
-          // Brackets: "[0][1]"
-          val brackets = chunk.substring(bracketIndex)
-          // Indices as a list: List(0, 1)
-          val listIndices = for {
-            m <- bracketMatcher.findAllMatchIn(brackets)
-            // It's possible for a bracket pair to be empty, in which case we just give it a random number
-            asInt = if (m.group(1).isEmpty) Random.nextInt() else m.group(1).toInt
-          } yield asInt
-          // Fold into a MetadataList: MetadataList(0 -> MetadataList(1 -> innerValue))
-          val metadataList = listIndices.toList.foldRight(innerValue)((index, acc) => MetadataList(TreeMap(index -> acc)))
-
-          MetadataObject(Map(objectName -> metadataList))
-        }
-    }
-  }
-
-  private def metadataValueToComponent(value: Option[MetadataValue], customOrdering: Option[Ordering[MetadataPrimitive]]): MetadataComponent = {
-    value map { someValue =>
-      val coerced: Try[MetadataPrimitive] = someValue.valueType match {
-        case MetadataInt => Try(MetadataPrimitive(JsNumber(someValue.value.toInt), customOrdering))
-        case MetadataNumber => Try(MetadataPrimitive(JsNumber(someValue.value.toDouble), customOrdering))
-        case MetadataBoolean => Try(MetadataPrimitive(JsBoolean(someValue.value.toBoolean), customOrdering))
-        case MetadataString => Try(MetadataPrimitive(JsString(someValue.value), customOrdering))
-      }
-
-      coerced match {
-        case Success(v) => v
-        case Failure(e) =>
-          log.warn(s"Failed to coerce ${someValue.value} to ${someValue.valueType}. Falling back to String.", e)
-          MetadataPrimitive(JsString(someValue.value), customOrdering)
-      }
-    } getOrElse MetadataEmpty
-  }
-
-  def customOrdering(event: MetadataEvent): Option[Ordering[MetadataPrimitive]] = event match {
-    case MetadataEvent(MetadataKey(_, Some(_), key), _, _) if key == CallMetadataKeys.ExecutionStatus => Option(MetadataPrimitive.ExecutionStatusOrdering)
-    case MetadataEvent(MetadataKey(_, None, key), _, _) if key == WorkflowMetadataKeys.Status => Option(MetadataPrimitive.WorkflowStateOrdering)
-    case _ => None
-  }
-  
-  private def toMetadataComponent(subWorkflowMetadata: Map[String, JsValue])(event: MetadataEvent) = {
-    lazy val originalKeyAndPrimitive = (event.key.key, metadataValueToComponent(event.value, customOrdering(event)))
-    
-    val keyAndPrimitive = if (event.key.key.endsWith(CallMetadataKeys.SubWorkflowId)) {
-      (for {
-        metadataValue <- event.value
-        subWorkflowMetadata <- subWorkflowMetadata.get(metadataValue.value)
-        keyWithSubWorkflowMetadata = event.key.key.replace(CallMetadataKeys.SubWorkflowId, CallMetadataKeys.SubWorkflowMetadata)
-        subWorkflowComponent = MetadataPrimitive(subWorkflowMetadata)
-      } yield (keyWithSubWorkflowMetadata, subWorkflowComponent)) getOrElse originalKeyAndPrimitive
-    } else originalKeyAndPrimitive
-
-    keyAndPrimitive._1.split(KeySeparator).foldRight(keyAndPrimitive._2)(parseKeyChunk)
-  }
 
   /**
     * Metadata for a call attempt
@@ -132,17 +54,8 @@ object MetadataBuilderActor {
     */
   private case class MetadataForIndex(index: Int, metadata: List[JsObject])
 
-  implicit val dateTimeOrdering: Ordering[OffsetDateTime] = scala.Ordering.fromLessThan(_ isBefore _)
-
-  /** Sort events by timestamp, transform them into TimestampedJsValues, and merge them together. */
-  private def eventsToIndexedJson(events: Seq[MetadataEvent], subWorkflowMetadata: Map[String, JsValue]): MetadataComponent = {
-    // The `List` has a `Foldable` instance defined in scope, and because the `List`'s elements have a `Monoid` instance
-    // defined in scope, `combineAll` can derive a sane `TimestampedJsValue` value even if the `List` of events is empty.
-    events.toList map toMetadataComponent(subWorkflowMetadata) combineAll
-  }
-
   private def eventsToAttemptMetadata(subWorkflowMetadata: Map[String, JsValue])(attempt: Int, events: Seq[MetadataEvent]) = {
-    val withAttemptField = JsObject(eventsToIndexedJson(events, subWorkflowMetadata).toJson.asJsObject.fields + (AttemptKey -> JsNumber(attempt)))
+    val withAttemptField = JsObject(MetadataComponent(events, subWorkflowMetadata).toJson.asJsObject.fields + (AttemptKey -> JsNumber(attempt)))
     MetadataForAttempt(attempt, withAttemptField)
   }
 
@@ -155,7 +68,7 @@ object MetadataBuilderActor {
   private def buildMetadataJson(events: Seq[MetadataEvent], includeCallsIfEmpty: Boolean, expandedValues: Map[String, JsValue]): JsObject = {
     // Partition events into workflow level and call level events
     val (workflowLevel, callLevel) = events partition { _.key.jobKey.isEmpty }
-    val workflowLevelJson = eventsToIndexedJson(workflowLevel, Map.empty).toJson.asJsObject
+    val workflowLevelJson = MetadataComponent(workflowLevel, Map.empty).toJson.asJsObject
 
     /*
      * Map(
@@ -192,11 +105,11 @@ object MetadataBuilderActor {
      * )
      */
     val callsGroupedByFQNAndIndexAndAttempt = callsGroupedByFQNAndIndex mapValues { _ mapValues { _ groupBy { _.key.jobKey.get.attempt } } }
-    
+
     val eventsToAttemptFunction = Function.tupled(eventsToAttemptMetadata(expandedValues) _)
     val attemptToIndexFunction = (attemptMetadataToIndexMetadata _).tupled
 
-    val callsMap = callsGroupedByFQNAndIndexAndAttempt mapValues { _ mapValues { _ map eventsToAttemptFunction } map attemptToIndexFunction } mapValues { md => 
+    val callsMap = callsGroupedByFQNAndIndexAndAttempt mapValues { _ mapValues { _ map eventsToAttemptFunction } map attemptToIndexFunction } mapValues { md =>
       JsArray(md.toVector.sortBy(_.index) flatMap { _.metadata })
     }
 
@@ -224,7 +137,7 @@ class MetadataBuilderActor(serviceRegistryActor: ActorRef) extends LoggingFSM[Me
 
   startWith(Idle, None)
   val tag = self.path.name
-  
+
   when(Idle) {
     case Event(action: MetadataServiceAction, _) =>
       serviceRegistryActor ! action
@@ -242,7 +155,7 @@ class MetadataBuilderActor(serviceRegistryActor: ActorRef) extends LoggingFSM[Me
     case Event(StatusLookupResponse(w, status), _) =>
       context.parent ! RequestComplete((StatusCodes.OK, processStatusResponse(w, status)))
       allDone
-    case Event(failure: ServiceRegistryFailure, _) =>
+    case Event(_: ServiceRegistryFailure, _) =>
       val response = APIResponse.fail(new RuntimeException("Can't find metadata service"))
       context.parent ! RequestComplete((StatusCodes.InternalServerError, response))
       allDone
@@ -270,25 +183,25 @@ class MetadataBuilderActor(serviceRegistryActor: ActorRef) extends LoggingFSM[Me
       context stop self
       stay()
   }
-  
+
   when(WaitingForSubWorkflows) {
     case Event(RequestComplete(metadata), Some(data)) =>
       processSubWorkflowMetadata(metadata, data)
   }
-  
+
   whenUnhandled {
     case Event(message, data) =>
       log.error(s"Received unexpected message $message in state $stateName with data $data")
       stay()
   }
-  
+
   def processSubWorkflowMetadata(metadataResponse: Any, data: MetadataBuilderActorData) = {
     metadataResponse match {
       case (StatusCodes.OK, js: JsObject) =>
         js.fields.get(WorkflowMetadataKeys.Id) match {
           case Some(subId: JsString) =>
             val newData = data.withSubWorkflow(subId.value, js)
-            
+
             if (newData.isComplete) {
               buildAndStop(data.originalQuery, data.originalEvents, newData.subWorkflowsMetadata)
             } else {
@@ -299,18 +212,18 @@ class MetadataBuilderActor(serviceRegistryActor: ActorRef) extends LoggingFSM[Me
       case _ => failAndDie(new RuntimeException("Failed to retrieve metadata for a sub workflow."))
     }
   }
-  
+
   def failAndDie(reason: Throwable) = {
     context.parent ! RequestComplete((StatusCodes.InternalServerError, APIResponse.error(reason)))
     context stop self
     stay()
   }
-  
+
   def buildAndStop(query: MetadataQuery, eventsList: Seq[MetadataEvent], expandedValues: Map[String, JsValue]) = {
     context.parent ! RequestComplete((StatusCodes.OK, processMetadataEvents(query, eventsList, expandedValues)))
     allDone
   }
-  
+
   def processMetadataResponse(query: MetadataQuery, eventsList: Seq[MetadataEvent]) = {
     if (query.expandSubWorkflows) {
       // Scan events for sub workflow ids