+++ supportedBackends/jes/src/main/scala/cromwell/backend/impl/jes/callcaching/JesBackendCacheHitCopyingActor.scala
@@ -48,16 +48,16 @@ class JesBackendCacheHitCopyingActor(standardParams: StandardCacheHitCopyingActo
       }
   }
 
-  override protected def additionalIoCommands(originalSimpletons: Seq[WdlValueSimpleton],
+  override protected def additionalIoCommands(sourceCallRootPath: Path,
+                                              originalSimpletons: Seq[WdlValueSimpleton],
                                               newOutputs: CallOutputs,
                                               originalDetritus:  Map[String, String],
                                               newDetritus: Map[String, Path]): List[Set[IoCommand[_]]] = {
-    val originalExecutionRoot = lookupSourceCallRootPath(originalDetritus)
     val content =
       s"""
          |This directory does not contain any output files because this job matched an identical job that was previously run, thus it was a cache-hit.
          |Cromwell is configured to not copy outputs during call caching. To change this, edit the filesystems.gcs.caching.duplication-strategy field in your backend configuration.
-         |The original outputs can be found at this location: ${originalExecutionRoot.pathAsString}
+         |The original outputs can be found at this location: ${sourceCallRootPath.pathAsString}
       """.stripMargin
 
     List(Set(writeCommand(jobPaths.callExecutionRoot / "call_caching_placeholder.txt", content, Seq(CloudStorageOptions.withMimeType("text/plain")))))